Completely ignore one-frame cuts
These seem to just be added to the project file to define a cut between two kept segments. I.e. if you "Add Selection" for 0->2s, and then again for 2s+1frame->3s, it does this between the 2s and 2s+1frame. So there's literally no point removing this as the output file seems to keep this frame.