Use project file's cut ms values instead of timecodes
The cut_start and cut_end values are Timecodes that I cannot figure out how to convert accurately to a full timestamp. The current Timestamp.from_timecode() method is as accurate as I can get it to be, but it isn't quite there yet.

The cut_time_start and cut_time_end values are milliseconds that can be converted to the accurate timestamps that I'm after.

Since FFmpeg and SubtitleEdit both expect Timestamp's rather than Timecodes, best to use the cut_time_* values rather than cut_*.